feat(kevent): DNS events (#184)
* Initial support for DNS events

* Implement DNS filter fields and extractor and create consumer tests for DNS events

* Remove debugging remnants

* Revert "Remove debugging remnants"

This reverts commit e2ee28200ac4240ada4b2d41fba8ddc926902f11.

* Remove local testing remnants

* Fix failing unit tests

+++ pkg/kevent/flags_test.go
@@ -22,7 +22,7 @@ import "testing"
 
 func TestParamFlags(t *testing.T) {
 	var tests = []struct {
-		flag     uint32
+		flag     uint64
 		flags    ParamFlags
 		expected string
 	}{
